🚚 Migrate `uses` to app directory
